+++ apps/server_app/main.cpp
@@ -1,13 +1,13 @@
 #include <QCoreApplication>
 #include <anticarium_pi/AnticariumManager.h>
-#include <anticarium_pi/config/ApplicationSettings.h>
+#include <anticarium_pi/Jttp.h>
 #include <spdlog/sinks/daily_file_sink.h>
 #include <spdlog/sinks/stdout_sinks.h>
 #include <spdlog/spdlog.h>
 
 int main(int argc, char* argv[]) {
     std::shared_ptr<spdlog::sinks::daily_file_sink_st> dailyLogger = std::make_shared<spdlog::sinks::daily_file_sink_st>("logs/AnticariumLog", 0, 0, false, 10);
-    std::shared_ptr<spdlog::sinks::stdout_sink_st> consoleLogger   = std::make_shared<spdlog::sinks::stdout_sink_st>();
+    std::shared_ptr<spdlog::sinks::stdout_sink_st> consoleLogger = std::make_shared<spdlog::sinks::stdout_sink_st>();
 
     std::vector<spdlog::sink_ptr> loggerSinks;
 
@@ -23,7 +23,7 @@ int main(int argc, char* argv[]) {
     SPDLOG_INFO("Program started");
 
     QCoreApplication a(argc, argv);
-    ApplicationSettings::instance(QCoreApplication::applicationDirPath(), QCoreApplication::instance());
+    JTTP::instance(QCoreApplication::instance());
     AnticariumManager anticariumManager;
     anticariumManager.run();