+++ cmake/ExternLibs.cmake
@@ -8,6 +8,11 @@ function(ADDQT5CORE TARGET_NAME)
 	target_link_libraries(${TARGET_NAME} PRIVATE Qt5::Core)
 endfunction()
 
+function(ADDQT5NETWORK TARGET_NAME)
+        find_package(Qt5 COMPONENTS Network REQUIRED)
+        target_link_libraries(${TARGET_NAME} PRIVATE Qt5::Network)
+endfunction()
+
 function(ADDJSON TARGET_NAME)
         find_package(nlohmann_json CONFIG REQUIRED)
         target_link_libraries(${TARGET_NAME} PRIVATE nlohmann_json nlohmann_json::nlohmann_json)
@@ -18,7 +23,7 @@ function(FETCHSHAREDTYPES)
 	FetchContent_Declare(
 	  shared_types
 	  GIT_REPOSITORY https://github.com/Anticarium/Shared_Types.git
-          GIT_TAG        v1.1.4
+          GIT_TAG        v1.3.0
 	)
 
 	FetchContent_GetProperties(shared_types)